TestApp/serializers.py
from rest_framework import serializers
from TestApp.models import Item

class ItemSerializer(serializers.ModelSerializer):
    class Meta:
        model = Item
        fields = ('id', 'name', 'description')