requirement.txt
asgiref==3.7.2
astor==0.8.1
attrdict==2.0.1
Babel==2.12.1
backports.zoneinfo==0.2.1
bce-python-sdk==0.8.83
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
contourpy==1.1.0
cssselect==1.2.0
cssutils==2.7.1
cycler==0.11.0
Cython==0.29.35
decorator==5.1.1
Django==4.2.2
djangorestframework==3.14.0
et-xmlfile==1.1.0
fire==0.5.0
Flask==2.3.2
flask-babel==3.1.0
fonttools==4.40.0
future==0.18.3
idna==3.4
imageio==2.31.1
imgaug==0.4.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
lazy-loader==0.2
lmdb==1.4.1
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
mod-wsgi==4.9.4
mysqlclient==2.1.1
networkx==3.1
numpy==1.24.3
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
paddle-bfloat==0.1.7
paddleocr==2.6.1.3
paddlepaddle==2.4.2
pandas==2.0.2
pdf2docx==0.5.6
Pillow==9.5.0
premailer==3.10.0
protobuf==3.20.0
psutil==5.9.5
pyclipper==1.3.0.post4
pycryptodome==3.18.0
PyMuPDF==1.20.2
pyparsing==3.1.0
python-dateutil==2.8.2
python-docx==0.8.11
pytz==2023.3
PyWavelets==1.4.1
rapidfuzz==3.1.1
rarfile==4.0
requests==2.31.0
scikit-image==0.21.0
scipy==1.10.1
shapely==2.0.1
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
termcolor==2.3.0
tifffile==2023.4.12
tqdm==4.65.0
typing-extensions==4.6.3
tzdata==2023.3
urllib3==2.0.3
visualdl==2.5.3
Werkzeug==2.3.6
zipp==3.15.0
